Added trivial tool filtering logic. (#1136)
This change adds support for `thv run --tools <name>` to only enable a subset of the tools available from an MCP server. The main focus of the implementation was to be self-contained so that we can swap it with something more appropriate at a later point in time.

Note that the original idea was to use Cedar to implement a policy at startup, but this added complexity for a piece of code that should eventually be swapped out.

Fixes #1003